Delegate api interaction to QCSObjects.
This commit does away with the QCSClient class and instead delegates
interaction with the different API endpoints to the QCSObjects themselves.

Adds tests for QCSObject.equivalent() method.

Testing led to discovery of in the constructor of NetworkProfiles that was setting the
ssh_port equal to the name, and this was fixed.

Due to the changes in how the API is accessed, the cleanup fixture in
`camayoc/tests/qcs/conftest.py` was able to be simplified.

All files using the QCSClient and QCSObjects were updated to use new tooling.

Also stubs out network profile CRUD test cases.

Closes #51
Closes #52