Test each fact from a scan on its own.
This commit separates the testing of each fact from testing whether or
not we can scan different profiles. So now there is a test parametrized
on profiles, scanning for each profile and confirming that is is
successful so far as it executes and generates a report, and a different
test parametrized on the facts generated from scans.