Add ability to test facts based on privilege level
Adds constants that contain what denial messages an underprivileged user
may encounter for given facts.

Now when testing facts, if it is a fact that's data varies based on
privilege level, we only test for an exact match to the test data for
privileged users, and otherwise test against the various denial messages
acceptable for that fact.

This commit also updates the gen_config.py script to be able to take in
a new column of data in the profiles csv indicating if the profile is
one with a privileged user.

Now no profiles are re-used from the original config, only profiles
indicated in the profiles csv are used.