Avoid extra scan by storing scan result as fact
This removes the separate test parametrized on the profiles that did a
whole other scan on the systems to record that test result. Instead, the
result of the scan is stored as a fact called "scan performed" and we
get a test result for that fact based on whether or not the scan got
through the whole process and produced a report.