Add QSC object for scan and one scan test
Adds a class to represent Scans that allows you to create scans, check on their
status, and alter status (pause/restart/cancel).

Includes one test that is parametrized on profiles found in the config file
that runs the scan and checks to see that it completes and generates a fact
collection.

Includes several other test stubs.

Closes #54