Enable satellite tests
Updates logic of some utility functions to allow for additional
"options" paramater that the satellite sources contain.

Also re-organizes where some utility functions are to prepare for report
tests that will need to use them as well.

Because only the "connect" type scan has been implemented at this time,
the tests for pausing and cancelling are set to "xfail" because until
"inspect" scans are implemented, we cannot test these functions.