Stub out tests for reports
Outline test cases for reports. Reports are still very much in flux and
we need some more details about what facts will be present, with what
priority they are chosen between sources when different information is
merged, and what queries to the reports endpoint will be acceptable.
That being said, this is a good start and gets the test cases in a
format so we can view them from our read the docs website.

Closes #95
Closes #94
Closes #98