Make API and CLI tests authenticate
Adds necessary code to api.Client to have the client send authenticated
requests if desired. This is enabled by default, but can be disabled
by calling api.Client().logout() or by passing a flag to the constructor
to disable it from the start. Also adds API login and logout tests.

Also has the CLI client login to the server after configuring the host
and port. This has the effect of all CLI tests being authenticated.

Negative tests for unauthenticated qpc commands and logout tests for the
CLI should be added at another time.

Closes #93
Closes #119