Enable scanning remote infrastructure with quipucords tests
Defines new section of `qcs` config to allow named scans of hosts to be run and
the reports analyized in tests. All scans are attempted at beginning of test
run and the scan id, report id, and other pertinent results are cached in a
global cache. Tests then run and can access this data and make whatever
assertions necessary.

Currenlty the tests are very minimal, mostly that the scan completed and some
results are available. This module would be a great place to flesh out for
product and os identification tests.

Also updates other tests to deal with config file change, and refactors
location of some code that was able to be re-used.

Closes #148