Test new delete workflow for credentials
Update existing tests and fixtures to deal with the fact that we must delete
all scans before we delete sources involved in them, and that we must delete
all sources before we delete credentials that depend on them.

Reworks how a couple tests work through paginated data from API because they
were failing to advance pages correctly.

Finally adds a test case confirming that we get denied if we attempt to delete
a credential if a source depends on it, and that we can then update the source
to use a new credential or we may delete the source and this then allows us to
delete the credential.

Closes #163