Add negative test case for disabled optional products
Test that the user cannot provide non-booleans to disable optional products.
It is the case that 'yes' and 'no' are valid booleans.

It is still the case that the server ignores extra input, this is a known
behavior, so we certainly can try and disable nonoptional products or
nonexistant ones, but the server will just ignore that input. This is not
considered a bug at this time.

Closes #171