Add scan job tests for disabled optional products
Creates constants for the facts collected by the optional product ansible roles.

Creates two tests that test the results of scans. One asserts that when
the optional products are disabled, the facts associated with the role that
scans for that products are not present in the report. The other test asserts
that the report correctly notes that the products were disabled for that scan.

Closes #193