Changes to allow for all scans to complete
Several changes to allow all the scans that happen at the beginning of the test
session to complete before the test suite proceeds. This includes allowing
longer boot times (but we proceed if the machine is responsive), and longer
wait times for a scan to complete since it may be queued behind several other
scans.

Additionally, we provide a fresh client to each test so that it has a current
authentication token and make sure the clients of objects about to be deleted
have a good token.

Closes #223