Don't wait as long for scans to start, pause, or cancel.
If all is well with the server, it should not take long for scans to start,
pause, or cancel because we are obtensibly waiting for all scans to complete
before starting a new test. If we do reach the timeout for these activities, it
is an indication that something is wrong.

Closes #242