Relax docs building requirement
Allow building docs even if there is warning issued.

Sphinx complain that
camayoc.ui.models.pages.credentials.NetworkCredentialForm.fill
contain reference to unknown target NetworkCredentialFormDTO. The problem
seems to be that fill method is decorated, and NetworkCredentialFormDTO
is type that is union of two other types. Our only conclusion is that
this setup is just too much for Sphinx, which is only trying to catch up
with Python type annotation changes.