Add end-to-end test for ocp source using cli route
Test was meant to be run after the proper provisioning of ocp instance,
which can be done with the script provision_ocp_instance.yaml. You would
need a running cluster with appropriate permissions to do so. Test
covers all the neccessary cli steps to perform a scan in an ocp
instance, including adding credential, source, scan, running it,
downloading the report and checking for its content. All entities will
be cleared by the end of it and file will be deleted.