Remove outdated tests
Disable optional products and Enable extended product search tests
assumed that certain facts will or will not be present in report. This
was true when tests were written in 2018, but changed in 2019 - now all
facts are always reported. Depending on these settings, certain tasks can
still be skipped, but this is only visible in server logs and can't be
easily asserted in Camayoc.

Tests in questions were removed from API. Tests in CLI are changed to
only check that reports with these settings can be created, run and
completed - as this provides a little coverage for qpc.