Force order of `camayoc/tests/qpc/cli/test_credentials.py::test_clear_all`
If this test is executed in the middle of a test run (where pytest puts
it by default), `qpc cred clear all` can't succeed - other tests already
created sources and scans depending on credentials.

As a workaround, force this test to be executed first. We have clean
database and we can create few credentials, remove them in batch and
expect removal to succeed.

This is a hack, but Camayoc is not prepared to run tests in random order
or in parallel, so it seems like an OK compromise.