Don't try to close wizard while UI is waiting for request response
At the final step of Source wizard, UI is making an actual request and
awaits response. This usually takes a split second. But when server is
particularly slow to respond, and Playwright is particularly fast, it
may try to close the wizard while request response is not yet received.
In this case, UI will display another dialog, asking for confirmation of
wizard termination.

This waits for request response to come back to us before closing the
wizard.