Stabilize test_scanjob_cancel
This test fails every few runs.

Sometimes it fails because job never starts - this is a known problem
with thread-based scan manager, and we can't do anything about this.

Sometimes it fails because we want scanjob to be running, but it
completed already. This seems to only happen for Ansible source. There's
nothing wrong with Ansible itself, but it seems that Ansible in our
environment is relatively small and tends to finish too quick. For now
let's ignore it as a source for this test.