Remove functions that were never implemented
While designing UI support in Camayoc, there were various ideas how
framework could work and what it could provide. Some of these ideas were
stubbed in empty functions and methods. A year later, it's clear that
these ideas are not going to be implemented - stubs should be removed to
not confuse people.