Run full UI end to end
As DISCOVERY-426 showed, sometimes removing scan that is still in progress
results in 500 error from server. This is rare (about 5% of tries) and a
bit of edge case (why would you try to delete a scan that is in
progress?), but may lowers our confidence in PR pipelines.

Since this never happens during standard (finished) scan removal, the
easiest way forward is to wait for scans to finish. And if we have to
wait for scans to finish, we might as well check in UI if we can
download them.

We wouldn't need to remove scans in the first place if it wasn't for
DISCOVERY-162. We have to keep number of credentials low, or we won't
be able to create new source.