Introduce nightly/pr_only markers
Builds foundation for DISCOVERY-628.

The idea is that you can mark tests to execute only in PR run or only in
nightly run. Or, consequently, to NOT execute in PR / nightly run. Tests
not marked with either are executed in both pipelines.

The actual implementation of marker-based selection is in discovery-ci.