Introduce UI tests focusing on scans
Introduce UI tests that take advantage of cached scans result and
download scans that already exist. They are very similar to existing end
to end tests.

The intent is that new scan tests are going to run only in PR check, to
speed things up. Old end to end tests are going to be run in nightly
pipeline, where execution speed is less of a concern. Since new scan
tests are very much included in old end to end tests, only one of them
should be executed in any given run.