test_scanjob_with_multiple_sources: test multiple source *types* again
Since the very beginning, the point of the test was to run a scan that
has different source *types*. I haven't noticed that and broke the test
in 78ac64636584854811e394947177035cdf91f887 - when test was ported to
data_provider and looked for a scan that has multiple sources in the
config. In our default config, there's only one scan like that, and it
uses multiple sources of *the same* type - network.