Introduce shared component for interacting with "Add ..." dropdown
When testing locally, occasionally we encounter the issue that clicking
"Add ..." button does not open dropdown (or dropdown disappears before
we can interact with it - impossible to tell). We need to wrap "click
add button and select item from dropdown" sequence in a loop, to retry
it. Since this is a component used in two places, we extract it to
shared component instead of repeating the loop.