New UI: Ensure there's single element to click
`.locator()` can match multiple elements on the page.
`.locator().click()` works in strict mode, i.e. Playwright will refuse
the operation if `.locator()` matches multiple elements. This is a
change compared to `.click(locator_string)`, which would happily operate
on first element matching the locator_string.

Using `.first` between `.locator()` and `.click()` ensures there's
single element to operate on.

This fixes a bug that is only present when running some tests on empty
database. Which happens to be a case in pipeline.