Introduce setting to not clean up on test suite run
Camayoc deletes all data it has created at the end of test suite run.
Historically it was because we used long-running Discovery servers. It
still provides value in implicitly testing deletion, and is still useful
in local runs when you are developing a test and want to start with
clean state every time.

But sometimes you want to have Discovery with some data, and you might
want to use Camayoc to create that data. And in this specific case you
don't want Camayoc to clean up because you are not really testing, you
are preparing the environment.

Thanks to this new setting, you can now decide if Camayoc should remove
data or not. The default is to remove.

Usually you will probably use it like that:

    DYNACONF_camayoc__db_cleanup=False pytest -v camayoc/tests/...