Move model from ssh_keyfile to ssh_key
In qpc_model representing credential, move from ssh_keyfile to ssh_key.
This also changes the meaning of the field - it used to represent path,
now it represents the content of the file.

However, when creating model from config definition, it still takes the
path and reads the content automatically. This way we don't have to
modify existing configuration files. And storing entire keys in
configuration files is going to be cumbersome.

However, at some point in the future we should probably streamline all
that and remove the cruft built around backwards compatibility and past
solutions.