Remove support for UI v1
This is semi-reversal of commits introduced in GitHub PR #543. All the
conditionals around `use_uiv2` setting are dropped. What used to be
inside conditional is now unversioned path, while what used to be v1
path is now removed (this applies mainly to field definitions and values).

This commit leaves around a bit of dead code that is going to be removed
in subsequent commits.