Introduce camayoc pytest plugin
The main goal is to introduce a new command line flag that allows pytest
to select correct tests for the pipeline. Normally tests selection would
be responsibility of a caller (in our case: discovery-ci), but pytest
does not expose all the information that caller needs to make kinds of
decisions that we want to make. So we are forced to implement actual
selection logic here in Camayoc.

We use pytest_addoption hook, which may be defined in standard
conftest.py file, but pytest is somewhat finicky when it comes to
supporting this specific hook from a file. Flag may or may not be
recognized depending on your current working directory and how you
invoke pytest. Registering a new plugin and defining hook there will
hopefully prove to be more robust solution.