Increase DBSnapshot robustness
In details report, a list of sources, and a list of facts within a
source, are fundamentally sets - order of items does not matter.

Order is consistent in API responses, but apparently it is affected by
database restoring / regenerate_missing_report_cache_files call. That's
why it was not noticed during initial implementation.