Try to prevent rare failure in UI tests
Login page has fade in animation set. We *think* extremely rare failures that
we observe on OpenStack might be caused by Playwright trying to insert
letters while animation is ongoing.

Let's wait about half a second before doing anything. If our assumption
about failure root cause is correct, it should help. If not - half a
second is small enough to disappear in normal variability (when testing
this locally, I had a run with delay that finished faster than a run
without this patch on).