Updated the credentials CLI tests to support --sshkeyfile
Updated use of the qpc CLI for sshkey, now specifying
the --sshkeyfile with the path of the SSH private key
to use or - for taking in the private key via stdin (the
current --sshkey behavior).

Added tests for specifying an SSH private key by file.
via --sshkeyfile <ssh_keyfile_path>