Add a test to verify Summary report in UI
A bit of explanation about `Client.page_contents` is in order.
UI framework is using fluent page model and all our UI tests follow a
pattern of single chain of method invocations, from starting a browser
to logging out. Reading data from page and running assertions on that is
a bit awkward. Especially when this is the first time where we actually
want to do such thing.

So the idea is that there are special methods that read something from a
browser and put that in `Client.page_contents`. After browser session
ends, you select requested data from `page_contents` and run assertions.
This way of working is consistent with how we handle downloads.

That approach will break when you want to change further actions in
browser session based on content obtained from a page, but as of this
commit there are no plans for such scenarios.

Relates to JIRA: DISCOVERY-982