Workaround an issue when we deselect item that is already selected
Sometimes after clicking "Credentials" input, one of items becomes
automatically selected. It's unclear why this is happening - it may be
related to Playwright. I have not observed it on OpenStack and I've seen
it happening in no more than 3% of local runs.

When it happens, our blind clicking on the element deselects it, instead
of selecting it. As a result, Credentials field becomes cleared and we fail
to submit data from modal.

That patch ensures only deselected elements are clicked.

Everything but a Network source allows only a single item to be selected.
Since we ignore already selected items, it may happen that we end up
trying to select item while something else is selected. That seems to
happen in about 0,1% of tests. We might try to deal with it if it
becomes a problem.