Introduce a test that checks if translations are correctly loaded
UI unit tests mock translations and never load them. Camayoc tests are
designed to work with page structure and elements with stable
identifiers. We end up in situation where no automation is checking if
translation file is valid and was correctly loaded.

Introduce a Camayoc test that does exactly that. The goal is not to
check if all the strings are correctly translated, but if any strings
are translated. This should catch low-probability high-impact event
where translations are not loaded at all.

Relates to JIRA: DISCOVERY-1140